The first part of WeatherController has been completed

+++ src/main/java/com/jean/servlet/model/owm/current/SysCurrentOWM.java
@@ -6,8 +6,8 @@ package com.jean.servlet.model.owm.current;
 public class SysCurrentOWM {
     private String message;
     private String country;
-    private int sunrise;
-    private int sunset;
+    private long sunrise;
+    private long sunset;
 
 
     public String getMessage() {
@@ -26,19 +26,19 @@ public class SysCurrentOWM {
         this.country = country;
     }
 
-    public int getSunrise() {
+    public long getSunrise() {
         return sunrise;
     }
 
-    public void setSunrise(int sunrise) {
+    public void setSunrise(long sunrise) {
         this.sunrise = sunrise;
     }
 
-    public int getSunset() {
+    public long getSunset() {
         return sunset;
     }
 
-    public void setSunset(int sunset) {
+    public void setSunset(long sunset) {
         this.sunset = sunset;
     }
 }
